+++ sound-mind-studio/tests/test_layer_controller.h
@@ -33,6 +33,7 @@ private slots:
     void addFilterLayerAddsAndSelectsAFilterLayer();
     void handleLayerSelectionChangedSyncsFilterConfigurationPanel();
     void applyFilterConfigurationAppliesOnlyToAFilterLayer();
+    void applyFilterConfigurationIsUndoableAndRedoable();
     void applyFilterConfigurationWithNoFilterLayerSelectedUpdatesThePendingConfiguration();
     void addFilterLayerSeedsFromThePendingFilterConfiguration();
     void setProjectResetsThePendingFilterConfiguration();