+++ sound-mind-studio/tests/test_macro_recorder.h
@@ -14,4 +14,6 @@ private slots:
     void startRecordingWhileAlreadyRecordingDoesNotClearEvents();
     void stopRecordingWhileNotRecordingIsANoOp();
     void recordingStateChangedEmitsOnStartAndStop();
+    void discardEventsClearsEventsAndStartUndoIndex();
+    void discardEventsWhileRecordingAlsoStopsRecording();
 };